GDScript: Split Function code into multiple files
To improve organization and reduce the size of compilation units.
